Fix deprecation of passing timeout as positional argument to brpop
The recent v4.8.0 of the redis gem deprecated passing `timeout` as a
positional argument in blocking commands like `brpop`:
https://github.com/redis/redis-rb/commit/08eea742485c054680ade3f9afd83f313c355595

v4.6 or greater of the redis gem is currently required and it already
included support for a `:timeout` hash option:
https://github.com/redis/redis-rb/blob/v4.6.0/lib/redis/distributed.rb#L465

Therefore we can switch to passing the timeout as a hash option.

Fixes #136.