Honor existing deployed limits
In a cluster with auto-scaling sidekiq servers, a new server entering
the pool runs Queues.start, which looks at the config file and sets
the limit and process_limit values accordingly.  However, if at
runtime these values have been changed dynamically for any reason,
this overwrites the values with what is in the config file, or removes
the limits if there are none in the config file.  This is unexpected
behavior.

Instead, at startup, honor the values found in redis if any are
present.  Only set the values if none are present in redis.